update version script
